django-backend/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
idna==3.4
install==1.3.5
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.3
